+++ .roadhogrc.mock.js
@@ -24,32 +24,38 @@ const proxy = {
     },
     $body: {
       name: 'Serati Ma',
-      avatar:
-        'https://gw.alipayobjects.com/zos/rmsportal/BiazfanxmamNRoxxVxka.png',
+      avatar: 'https://gw.alipayobjects.com/zos/rmsportal/BiazfanxmamNRoxxVxka.png',
       userid: '00000001',
       email: 'antdesign@alipay.com',
       signature: '海纳百川，有容乃大',
       title: '交互专家',
       group: '蚂蚁金服－某某某事业群－某某平台部－某某技术部－UED',
-      tags: [{
-        key: '0',
-        label: '很有想法的',
-      }, {
-        key: '1',
-        label: '专注设计',
-      }, {
-        key: '2',
-        label: '辣~',
-      }, {
-        key: '3',
-        label: '大长腿',
-      }, {
-        key: '4',
-        label: '川妹子',
-      }, {
-        key: '5',
-        label: '海纳百川',
-      }],
+      tags: [
+        {
+          key: '0',
+          label: '很有想法的',
+        },
+        {
+          key: '1',
+          label: '专注设计',
+        },
+        {
+          key: '2',
+          label: '辣~',
+        },
+        {
+          key: '3',
+          label: '大长腿',
+        },
+        {
+          key: '4',
+          label: '川妹子',
+        },
+        {
+          key: '5',
+          label: '海纳百川',
+        },
+      ],
       notifyCount: 12,
       country: 'China',
       geographic: {
@@ -176,6 +182,176 @@ const proxy = {
   },
   'GET /api/geographic/province': getProvince,
   'GET /api/geographic/city/:province': getCity,
+  'GET /api/menus': [
+    {
+      name: 'dashboard',
+      icon: 'dashboard',
+      path: 'dashboard',
+      children: [
+        {
+          name: '分析页',
+          path: 'analysis',
+        },
+        {
+          name: '监控页',
+          path: 'monitor',
+        },
+        {
+          name: '工作台',
+          path: 'workplace',
+          // hideInBreadcrumb: true,
+          // hideInMenu: true,
+        },
+      ],
+    },
+    {
+      name: '表单页',
+      icon: 'form',
+      path: 'form',
+      children: [
+        {
+          name: '基础表单',
+          path: 'basic-form',
+        },
+        {
+          name: '分步表单',
+          path: 'step-form',
+        },
+        {
+          name: '高级表单',
+          authority: 'admin',
+          path: 'advanced-form',
+        },
+      ],
+    },
+    {
+      name: '列表页',
+      icon: 'table',
+      path: 'list',
+      children: [
+        {
+          name: '查询表格',
+          path: 'table-list',
+        },
+        {
+          name: '标准列表',
+          path: 'basic-list',
+        },
+        {
+          name: '卡片列表',
+          path: 'card-list',
+        },
+        {
+          name: '搜索列表',
+          path: 'search',
+          children: [
+            {
+              name: '搜索列表（文章）',
+              path: 'articles',
+            },
+            {
+              name: '搜索列表（项目）',
+              path: 'projects',
+            },
+            {
+              name: '搜索列表（应用）',
+              path: 'applications',
+            },
+          ],
+        },
+      ],
+    },
+    {
+      name: '详情页',
+      icon: 'profile',
+      path: 'profile',
+      children: [
+        {
+          name: '基础详情页',
+          path: 'basic',
+        },
+        {
+          name: '高级详情页',
+          path: 'advanced',
+          authority: 'admin',
+        },
+      ],
+    },
+    {
+      name: '结果页',
+      icon: 'check-circle-o',
+      path: 'result',
+      children: [
+        {
+          name: '成功',
+          path: 'success',
+        },
+        {
+          name: '失败',
+          path: 'fail',
+        },
+      ],
+    },
+    {
+      name: '异常页',
+      icon: 'warning',
+      path: 'exception',
+      children: [
+        {
+          name: '403',
+          path: '403',
+        },
+        {
+          name: '404',
+          path: '404',
+        },
+        {
+          name: '500',
+          path: '500',
+        },
+        {
+          name: '触发异常',
+          path: 'trigger',
+          hideInMenu: true,
+        },
+      ],
+    },
+    {
+      name: '账户',
+      icon: 'user',
+      path: 'user',
+      authority: 'guest',
+      children: [
+        {
+          name: '登录',
+          path: 'login',
+        },
+        {
+          name: '注册',
+          path: 'register',
+        },
+        {
+          name: '注册结果',
+          path: 'register-result',
+        },
+      ],
+    },
+    {
+      name: '个人页',
+      icon: 'user',
+      path: 'account',
+      children: [
+        {
+          name: '个人中心',
+          path: 'center',
+        },
+        {
+          name: '个人设置',
+          path: 'settings',
+        },
+      ],
+    },
+  ],
 };
 
 export default (noProxy ? {} : delay(proxy, 1000));
