+++ src/router.js
@@ -3,7 +3,7 @@ import { routerRedux, Route, Switch } from 'dva/router';
 import { LocaleProvider, Spin } from 'antd';
 import zhCN from 'antd/lib/locale-provider/zh_CN';
 import dynamic from 'dva/dynamic';
-import { getRouterData } from './common/router';
+import { getRouterConfig } from './common/router';
 import Authorized from './utils/Authorized';
 import styles from './index.less';
 
@@ -14,7 +14,7 @@ dynamic.setDefaultLoadingComponent(() => {
 });
 
 function RouterConfig({ history, app }) {
-  const routerData = getRouterData(app);
+  const routerData = getRouterConfig(app);
   const UserLayout = routerData['/user'].component;
   const BasicLayout = routerData['/'].component;
   return (
