+++ src/services/auth.js
@@ -1,0 +1,5 @@
+import request from '../utils/request';
+
+export async function getAuthMenus() {
+  return request('/api/menus');
+}
